+++ js-miniapp-bridge/src/types/error-types/index.ts
@@ -46,17 +46,11 @@ function parseMiniAppError(jsonString: string): MiniAppError {
       parseSimError(json) ||
       new MiniAppError(json)
     );
-  } catch (e) {
-    console.error(e);
-    if (jsonString !== '' || jsonString !== undefined) {
-      return new MiniAppError({
-        type: 'MiniAppError',
-        message: jsonString,
-      });
-    }
+  } catch (_e) {
+    console.error(_e);
     return new MiniAppError({
       type: 'MiniAppError',
-      message: 'Failed to parse the error',
+      message: jsonString || 'Failed to parse the error',
     });
   }
 }